+++ package.json
@@ -30,6 +30,7 @@
     "gzip-js": "~0.3.2",
     "jest-cli": "^0.6.1",
     "platform": "^1.1.0",
+    "react-current-owner": "file:packages/react-current-owner",
     "run-sequence": "^1.1.4",
     "through2": "^2.0.0",
     "tmp": "~0.0.28",
