+++ packages/react-current-owner/README.md
@@ -1,0 +1,4 @@
+# `react-current-owner`
+
+This package serves as the shared CurrentOwner of all React renderer implementations. It should not be relied upon directly.
+
