+++ packages/react-current-owner/index.js
@@ -5,12 +5,25 @@
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
- *
- * @providesModule ReactCurrentOwner
  */
 
 'use strict';
 
-var ReactCurrentOwner = require('react-current-owner');
+/**
+ * Keeps track of the current owner.
+ *
+ * The current owner is the component who should own any components that are
+ * currently being constructed.
+ */
+var ReactCurrentOwner = {
+
+  /**
+   * @internal
+   * @type {ReactComponent}
+   */
+  current: null,
+
+};
 
 module.exports = ReactCurrentOwner;
+
