--- packages/react/lib/React.native.js
@@ -1,5 +1,0 @@
-'use strict';
-
-// TODO: Once we remove the DOM bits from React, this shim can go away.
-
-module.exports = require('./ReactIsomorphic');
