+++ src/ReactBrowser.js
@@ -1,0 +1,33 @@
+/**
+ * Copyright 2013-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+'use strict';
+
+/**
+ * The React Browser distribution really has ReactDOM bundled in side of it
+ * where the NPM distros are fully decoupled.
+ *
+ * TODO: decouple the browser builds as well.
+ */
+var ReactDOM = require('ReactDOM');
+var ReactDOMServer = require('ReactDOMServer');
+var ReactIsomorphic = require('ReactIsomorphic');
+
+var assign = require('Object.assign');
+
+// `version` will be added here by ReactIsomorphic.
+var React = {};
+
+assign(React, ReactIsomorphic);
+
+React.__SECRET_DOM_DO_NOT_USE_OR_YOU_WILL_BE_FIRED = ReactDOM;
+React.__SECRET_DOM_SERVER_DO_NOT_USE_OR_YOU_WILL_BE_FIRED = ReactDOMServer;
+
+module.exports = React;
+
